requirements.txt
black==24.2.0
docx2txt==0.8
faiss-cpu==1.7.4
httpx==0.27.0
isort==5.13.2
jupyter==1.0.0
langchain-community==0.0.19
langchain-experimental==0.0.51
langchain-openai==0.0.5
langchain==0.1.6
loguru==0.7.2
notebook==7.1.0
openai-whisper==20231117
openai==1.11.1
pandas==2.2.0
pydub==0.25.1
PyMuPDF==1.23.21
python-dotenv==1.0.1
sentence-transformers==2.2.0  # Added the version for sentence_transformers
streamlit==1.31.1
tiktoken==0.5.2
watchdog==4.0.0
selenium==4.4.3  # Added Selenium for web scraping
beautifulsoup4==4.11.1  # Added BeautifulSoup for parsing HTML
chromedriver-binary==104.0.5112.79.0  # Added chromedriver-binary for convenience, adjust the version as needed
concurrent-log-handler==0.9.19  # For handling logging in a concurrent environment
plotly==5.19.0
mysql-connector-python==8.3.0
tqdm==4.66.2
praw==7.7.1
matplotlib==3.8.3